move dbutil, httputil, logger into util folder
