+++ cmd/proxy/proxy.go
@@ -14,7 +14,7 @@ import (
 	"github.com/google/gops/agent"
 	"github.com/skycoin/skycoin/src/cipher"
 	"github.com/skycoin/teller/src/daemon"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/logger"
 	"github.com/skycoin/teller/src/proxy"
 )
