+++ src/proxy/gateway.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 
 	"github.com/skycoin/teller/src/daemon"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/logger"
 )
 
 type gatewayer interface {
