+++ src/service/exchange/exchange.go
@@ -15,7 +15,7 @@ import (
 
 	"github.com/skycoin/teller/src/daemon"
 	"github.com/skycoin/teller/src/util/dbutil"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/logger"
 	"github.com/skycoin/teller/src/service/scanner"
 	"github.com/skycoin/teller/src/service/sender"
 )
