+++ src/service/handlers_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/skycoin/teller/src/daemon"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/logger"
 	"github.com/stretchr/testify/require"
 )
