+++ src/service/monitor/monitor.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/skycoin/teller/src/daemon"
 	"github.com/skycoin/teller/src/util/httputil"
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/logger"
 	"github.com/skycoin/teller/src/service/exchange"
 )
