+++ src/service/sender/send_service.go
@@ -11,7 +11,7 @@ import (
 	"github.com/skycoin/skycoin/src/api/webrpc"
 	"github.com/skycoin/skycoin/src/cipher"
 
-	"github.com/skycoin/teller/src/logger"
+	"github.com/skycoin/teller/src/util/logger"
 )
 
 const sendCoinCheckTime = 3 * time.Second
